requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2
django-allauth==0.42.0
django-cors-headers==3.7.0
django-countries==7.1
django-filter==2.4.0
django-heroku==0.3.1
django-restframework==0.0.1
django-six==1.0.4
djangorestframework==3.12.4
idna==2.10
Markdown==3.3.4
oauthlib==3.1.0
Pillow==8.2.0
psycopg2==2.8.6
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
validate-email==1.3
whitenoise==5.2.0
